requirements.txt
pydantic~=2.10.2
fastapi~=0.111.1
uvicorn~=0.30.1
opencv-python~=4.9.0.80
numpy~=1.26.4
langchain~=0.3.12
langchain_community~=0.3.12
unstructured~=0.16.11
pi_heif~=0.21.0
unstructured_inference~=0.8.1
pytesseract~=0.3.13
unstructured_pytesseract~=0.3.13
langchain_huggingface~=0.1.2
google-generativeai
loguru~=0.7.3
faiss-cpu
openai-clip
